requirements-dev.txt
asgiref==3.9.1
asttokens==3.0.0
black==25.9.0
click==8.3.0
crispy-tailwind==1.0.3
decorator==5.2.1
Django==5.2.6
django-crispy-forms==2.4
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.1
executing==2.2.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Markdown==3.9
matplotlib-inline==0.1.7
mypy_extensions==1.1.0
packaging==25.0
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pytokens==0.1.10
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.2.13
